requirements.txt
pillow-heif>=0.20.0
Pillow>=10.1.0